cleaned up premake files and fixed all warnings except for 1
API entry macro redefinition, looks like it's caused by glad and/or glfw?

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -6,7 +6,6 @@
 #include "examples/imgui_impl_glfw.h"
 #include "examples/imgui_impl_opengl3.h"
 
-#include <glad/glad.h>
 #include <GLFW/glfw3.h>
 
 namespace Hazel {
@@ -61,7 +60,7 @@ namespace Hazel {
 	void ImGuiLayer::End() {
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
